create separate js for media management
